src/agent-framework/pyproject.toml
[project]
name = "agent-framework"
version = "0.1.0"
description = "A modular agent framework based on LangChain"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "pypdf>=4.0.0",
    "docx2txt>=0.8",
    "mysql-connector-python>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
